WIP main programme calls

+++ Mortgage_Calculator/userInput.cs
@@ -4,8 +4,8 @@ namespace Mortgage_Calculator
     public class UserInput
     {
 
-        private int _amount;
-        public int Amount
+        private double _amount;
+        public double Amount
         {
             get { return _amount; }
             set { _amount = value; }
